refactor(downloader): extract download_pre_resolved_fragments to shared module
Mechanical move of DashDownloader's private download_pre_resolved_fragments
(plus resolve_fragment_url / fetch_fragment_bytes free helpers) into a new
rdlp-downloader::fragments module. DashDownloader::download_format now calls
the shared free function. No behavior change.

Prepares for HLS reuse (issue #245).
